+++ docs_sources/releases.md
@@ -1,6 +1,6 @@
 # scikit-rebate 0.3
 
-* Added a parallelization parameter, `n_jobs`, to ReliefF, SURF, SURF*, and MultiSURF via joblib. This feature is currently only supported in Python 3+.
+* Added a parallelization parameter, `n_jobs`, to ReliefF, SURF, SURF*, and MultiSURF via joblib.
 
 * Renamed the `dlimit` parameter to `discrete_limit` to better reflect the purpose of the parameter.